Move harris corner detection from filter to feature subpackage.
