+++ skimage/feature/__init__.py
@@ -1,3 +1,4 @@
 from hog import hog
 from greycomatrix import greycomatrix, greycoprops
 from peak import peak_local_max
+from harris import harris
