+++ skimage/filter/__init__.py
@@ -5,4 +5,3 @@ from edges import sobel, hsobel, vsobel, hprewitt, vprewitt, prewitt
 from tv_denoise import tv_denoise
 from rank_order import rank_order
 from thresholding import threshold_otsu
-from harris import harris
